requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.0.6
django-extensions==3.2.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
drf-flex-fields==0.9.8
drf-nested-routers==0.93.4
drf-spectacular==0.23.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
jsonschema==4.14.0
msgpack==1.0.4
mysqlclient==2.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
zope.interface==5.4.0
